+++ bpfprogs/bpfdebug.go
@@ -28,7 +28,7 @@ func SetupBPFDebug(ebpfChainDebugAddr string, BPFConfigs *NFConfigs) {
 			}
 			listener, err := net.ListenTCP("tcp", tcpAddr)
 			if err != nil {
-				log.Fatal().Err(err).Msgf("Not able to create net Listen")
+				log.Fatal().Err(err).Msgf("unable to create net Listen")
 			}
 			models.AllNetListeners.Store("debug_http", listener)
 		}
